refactor: Centrelize tracks functions in MatomoDrive File
