+++ app/src/main/java/com/infomaniak/drive/ui/fileList/multiSelect/MultiSelectFragment.kt
@@ -33,7 +33,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.RecyclerView
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.google.android.material.appbar.CollapsingToolbarLayout
-import com.infomaniak.drive.MatomoDrive.trackEvent
+import com.infomaniak.drive.MatomoDrive.trackBulkActionEvent
 import com.infomaniak.drive.R
 import com.infomaniak.drive.data.api.UploadTask.Companion.LIMIT_EXCEEDED_ERROR_CODE
 import com.infomaniak.drive.data.models.BulkOperation
@@ -57,7 +57,6 @@ import com.infomaniak.drive.utils.Utils.duplicateFilesClicked
 import com.infomaniak.drive.utils.Utils.moveFileClicked
 import com.infomaniak.drive.utils.showSnackbar
 import com.infomaniak.lib.core.utils.ApiErrorCode.Companion.translateError
-import com.infomaniak.lib.core.utils.capitalizeFirstChar
 import com.infomaniak.lib.core.utils.whenResultIsOk
 
 abstract class MultiSelectFragment(private val matomoCategory: String) : Fragment(), MultiSelectResult {
@@ -521,14 +520,6 @@ abstract class MultiSelectFragment(private val matomoCategory: String) : Fragmen
         closeMultiSelect()
 
         onAllIndividualActionsFinished(type)
-    }
-
-    private fun trackBulkActionEvent(category: String, action: BulkOperationType, modifiedFileNumber: Int) {
-
-        fun BulkOperationType.toMatomoString(): String = name.lowercase().capitalizeFirstChar()
-
-        val name = "bulk" + (if (modifiedFileNumber == 1) "Single" else "") + action.toMatomoString()
-        trackEvent(category, name, value = modifiedFileNumber.toFloat())
     }
 
     companion object {
